+++ src/Extension/ExtHelpers.h
@@ -1,0 +1,5 @@
+#pragma once
+
+extern  ELanguageIndicatorMode iShowLanguageInTitle;
+
+
